Restructure project
Restructure project to absorb the recently merged auxiliary repos as
libraries in the taskw package.

The taskd-setup repo is now the taskw/taskd library. For faster
iteration on development, the pem files are now hard-coded. The
library was refactored so that unit tests might be easier to
write. The CI tests no longer need to use task, taskd or taskd-setup
executables, aside from implicitly in `dart test`.

The taskd-client repo is now the taskw/taskc and taskw/taskc_impl
libraries. Codec, message and send were moved to taskw/taskc_impl, as
implementation details, so that the taskw/taskc API documentation
generated by dartdoc might be slightly easier to understand. Still
very few docstrings though, only those for library level files.

Json objects were moved to taskw/json, with plans to improve how they
are split between what is needed for storing, syncing or exporting
tasks.

The parseTaskrc method was moved to the small taskw/taskrc library,
with a vague intention of coming up with more options for
configuration.